improved text search replacing "." and "," for " "
Added 2 lines that change the new_text input that replaces "." and "," to " ", that makes it better for searching words with the extra space after the "word " so it knows better where is the end of the word ! It wont mistake "safety" for "safe" It is far from perfect. But hey, give it a try and tell me something interesting about it!
Just Install Godot Engine first, it's free for Windows Mac and Linux It is amazing ! :)
Download the file, unzip it and import the directory into Godot. Open it and press F5.
Write help If you need any