Improved search Engine and Responses!
Added an " " extra space before the text input and before every search for word. Also replacing "!" for " ".
Now it wont mistake " used " with " confused ". 
Did a Few more changes to the text and code cleaning 
ahhh there is a face to to the Chat bot now an she is blond as my memory of Angelica Rente.